master : menambahkan invoice.php di controler ddan menambahkan model_invoice.php di model

+++ application/controllers/Dashboard.php
@@ -51,12 +51,15 @@ class Dashboard extends CI_Controller{
     
     public function proses_pesanan()
     {
-        {
+        $is_processed = $this->model_invoice->index();
+        if($is_processed){
             $this->cart->destroy();
             $this->load->view('templates/header');
             $this->load->view('templates/sidebar');
             $this->load->view('proses_pesanan');
             $this->load->view('templates/footer');
-        }   
+        }else{
+            echo "Maaf, Pesanan Anda Gagal diproses! ";
+        }  
     }
 }
